Remvoe the old implementation and '2' labeling
